create interface to document
